+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add pipe command `SET_TRIAL_ID` to allow to annotate data samples.
 - On multi-screen setups start calibration with a screen selection page.
 - Add Screen Area coordinates to the dumped configuration file.
+- Add gaze validation which can be started through the command `VALIDATE`.
 
 ### Improvements
 - Fix compiler warnings.